+++ src/potent/operations/_base.py
@@ -181,3 +181,20 @@ class BaseOperation(CommonBase):
         # return fields
 
         return lines
+
+    @classmethod
+    def to_markdown_summary(cls) -> tuple[str, str]:
+        """
+        Returns a tuple of (slug, required_config) to summarize this operation. The slug is also a link to the
+        """
+
+        fields = cls.model_fields
+
+        config_required = False
+        if (raw_config := fields.get("config")) and raw_config.annotation:
+            config_required = True
+
+        return (
+            f"[`{get_args(fields['slug'].annotation)[0]}`](#{cls.__name__})",
+            "☑️" if config_required else "",
+        )