+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- ‼️ change the `switch-branch` slug to `git-switch` to be more in line with other commands
 - Add the `--version` cli flag to the root CLI command
 
 ## 0.2.2